+++ src/main/java/com/zea7ot/leetcode/logs.txt
@@ -24,7 +24,7 @@
 5: 
 
 20201019:
-3: 0289, 0379, 0609, 1236
+3: 0289, 0379, 0547, 0609, 1236
 4: 0486
 5: 0291
