+++ src/main/java/com/zea7ot/leetcode/lvl3/lc0547/SolutionApproach0DFS.java
@@ -1,0 +1,16 @@
+/**
+ * https://leetcode.com/problems/friend-circles/
+ * 
+ * Time Complexity:     O(`NR` * `NC`)
+ * Space Complexity:    O(`NR` * `NC`)
+ * 
+ * References:
+ *  https://leetcode.com/problems/friend-circles/discuss/101431/Stupid-question:-How-is-this-question-different-from-Number-of-Islands/332347
+ *  https://leetcode.com/problems/friend-circles/discuss/101338/Neat-DFS-java-solution/105148
+ *  https://leetcode.com/problems/friend-circles/discuss/101338/Neat-DFS-java-solution
+ */
+package com.zea7ot.leetcode.lvl3.lc0547;
+
+public class SolutionApproach0DFS {
+
+}
