+++ src/main/java/com/zea7ot/leetcode/tags/by_algorithm/sort/bucket_sort.txt
@@ -6,6 +6,7 @@ References:
 ### Sort
 # Bucket Sort
 
+2   0362    https://leetcode.com/problems/design-hit-counter/
 2   0791    https://leetcode.com/problems/custom-sort-string/
 2   1365    https://leetcode.com/problems/how-many-numbers-are-smaller-than-the-current-number/
