+++ src/main/java/com/zea7ot/leetcode/tags/by_algorithm/sort/summary.txt
@@ -9,6 +9,7 @@ References:
 
 2   0215    https://leetcode.com/problems/kth-largest-element-in-an-array/
 2   0280    https://leetcode.com/problems/wiggle-sort/
+2   0362    https://leetcode.com/problems/design-hit-counter/
 2   0451    https://leetcode.com/problems/sort-characters-by-frequency/
 2   0791    https://leetcode.com/problems/custom-sort-string/
 2   1029    https://leetcode.com/problems/two-city-scheduling/
