+++ src/main/java/com/zea7ot/leetcode/tags/by_data_structure/array/by_data_structure/array_1d_circular.txt
@@ -6,6 +6,7 @@ References:
 ### Array
 # Circular Array
 
+2   0362    https://leetcode.com/problems/design-hit-counter/
 2   0622    https://leetcode.com/problems/design-circular-queue/
 2   0933    https://leetcode.com/problems/number-of-recent-calls/
