+++ src/main/java/com/zea7ot/leetcode/tags/by_data_structure/array/by_dimension/array_1d.txt
@@ -27,6 +27,7 @@ References:
 2   0278    https://leetcode.com/problems/first-bad-version/
 2   0280    https://leetcode.com/problems/wiggle-sort/
 2   0303    https://leetcode.com/problems/range-sum-query-immutable/
+2   0362    https://leetcode.com/problems/design-hit-counter/
 2   0451    https://leetcode.com/problems/sort-characters-by-frequency/
 2   0605    https://leetcode.com/problems/can-place-flowers/
 2   0704    https://leetcode.com/problems/binary-search/
