+++ src/main/java/com/zea7ot/leetcode/tags/by_data_structure/array/summary.txt
@@ -30,6 +30,7 @@ This file serves as the total list of problems related to Array, regardless of a
 2   0280    https://leetcode.com/problems/wiggle-sort/
 2   0303    https://leetcode.com/problems/range-sum-query-immutable/
 2   0310    https://leetcode.com/problems/minimum-height-trees/
+2   0362    https://leetcode.com/problems/design-hit-counter/
 2   0451    https://leetcode.com/problems/sort-characters-by-frequency/
 2   0463    https://leetcode.com/problems/island-perimeter/
 2   0485    https://leetcode.com/problems/max-consecutive-ones/
