+++ src/main/java/com/zea7ot/leetcode/tags/multithread.txt
@@ -5,4 +5,6 @@ References:
 
 ### Multi-Thread
 
-1242    https://leetcode.com/problems/web-crawler-multithreaded/
+2   0362    https://leetcode.com/problems/design-hit-counter/
+
+?   1242    https://leetcode.com/problems/web-crawler-multithreaded/
