+++ src/main/java/com/zea7ot/leetcode/tags/ood.txt
@@ -7,6 +7,7 @@ References:
 ### Object-Oriented Design
 
 2   0359    https://leetcode.com/problems/logger-rate-limiter/
+2   0362    https://leetcode.com/problems/design-hit-counter/
 2   0384    https://leetcode.com/problems/shuffle-an-array/
 2   0398    https://leetcode.com/problems/random-pick-index/
 2   0535    https://leetcode.com/problems/encode-and-decode-tinyurl/
